Allowed errors should depend on number of clients.
The default behavior of the federated client
is to only emit an error if all clients fail.
However, "all clients" should be the total number of clients,
not the number of clients that are currently fetching a pattern.
E.g., if there are 3 clients, but only 1 of them can have a match
(other matches have been ruled out by an optimization),
then no error should be emitted if that 1 client fails.