Emit parser errors on fragment.
The page is a source of the parser iterator,
so any error emitted on the page (source) would be
re-emitted on the parser (destination) by AsyncIterator.
Therefore, an endless loop occurs
if the parser in turn bounces errors back to the page.
Instead, the parser should emit them on the fragment,
because the fragment reads from the parser.